rename folder && init browser.ts
