+++ src/code/browser/workbench/editor/titleBar/titleBar.ts
@@ -1,7 +1,7 @@
-import { Component } from 'src/code/workbench/browser/component';
-import { WindowBarComponent } from 'src/code/workbench/browser/editor/titleBar/windowBar';
-import { FunctionBarComponent } from 'src/code/workbench/browser/editor/titleBar/functionBar';
-import { EditorComponentType } from 'src/code/workbench/browser/editor/editor';
+import { Component } from 'src/code/browser/workbench/component';
+import { WindowBarComponent } from 'src/code/browser/workbench/editor/titleBar/windowBar';
+import { FunctionBarComponent } from 'src/code/browser/workbench/editor/titleBar/functionBar';
+import { EditorComponentType } from 'src/code/browser/workbench/editor/editor';
 
 /**
  * @description TitleBarComponent stores and handles all the titleBar and functionBar
